+++ src/components/SInput/SInput.stories.ts
@@ -189,6 +189,16 @@ export const WithSuffix = createVariation(`
 <SInput suffix="optional" />
 `);
 
+export const WithOptionsEmbedded = createVariation(`
+<SInput>
+  <template #options :name="currency">
+    <option value="usd">USD</option>
+    <option value="cop">COP</option>
+    <option value="eur">EUR</option>
+  </template>
+</SInput>
+`);
+
 /* import {
   SInput,
   SInputDropdown,